feat: documented
