+++ memory-bank/reflection/reflection-workspace-migration.md
@@ -1,0 +1,88 @@
+# Task Reflection: Migrate Bundled Plugins to Workspace Interface
+
+**Feature ID:** workspace-migration
+**Date of Reflection:** 2026-02-15
+**Complexity Level:** 3 (Intermediate Feature)
+
+## Brief Feature Summary
+
+Replaced all direct `fs.*` calls (~72 total) across the three bundled plugins (plugin-a16n, plugin-claude, plugin-cursor) with `Workspace` interface methods. This completes the workspace abstraction started in the previous architectural redesign, enabling `MemoryWorkspace`-based testing of plugin internals.
+
+## 1. Overall Outcome & Requirements Alignment
+
+- **All 5 strategic goals met:**
+  1. All `fs.*` calls in plugins replaced with `Workspace` methods
+  2. `LocalWorkspace` moved to `@a16njs/models` (plugins can't depend on engine)
+  3. Shared utilities (`readAgentSkillIO`/`writeAgentSkillIO`) bridged via `ws.resolve()`
+  4. 100% backward compatibility maintained — `string` args still work
+  5. Zero regressions — 807 tests pass (up from 804 baseline due to new tests)
+- **No scope deviations.** The plan was followed exactly as written across all 5 phases.
+- **Overall assessment:** Fully successful. Clean mechanical migration with no surprises.
+
+## 2. Planning Phase Review
+
+- **Plan accuracy:** Excellent. The fs call audit (72+ calls) and per-plugin breakdown were accurate. The phase ordering (models infrastructure first, then smallest-to-largest plugin) proved effective.
+- **Workspace method mapping table** in the plan was immediately actionable — every `fs` call mapped 1:1 to a workspace method.
+- **Risk assessment was accurate:** The identified risks (recursive traversal differences, WorkspaceEntry vs Dirent, path normalization) all materialized as non-issues because the plan accounted for them.
+- **What could have been planned better:** The plan could have noted that `readAgentSkillIO`/`writeAgentSkillIO` would keep string-based APIs (bridged via `ws.resolve()`) rather than being migrated to accept `Workspace`. This was a reasonable decision made during Phase 1 but wasn't pre-planned.
+
+## 3. Creative Phase(s) Review
+
+No creative phases were required — this was a mechanical migration with clear mapping from old API to new. The right call.
+
+## 4. Implementation Phase Review
+
+### Major Successes
+- **Consistent migration pattern:** Once established in plugin-a16n (Phase 1), the pattern (`toWorkspace()` at entry, `ws.*` methods in helpers, boolean property access for `WorkspaceEntry`) applied identically to plugin-claude and plugin-cursor with zero adaptation needed.
+- **Test reuse:** Existing tests for plugin-claude and plugin-cursor required zero modifications because the public API (`string | Workspace`) still accepts strings. Only plugin-a16n's `parseIRFile` tests needed updating (internal function with changed signature).
+- **Phase gates worked:** Building and testing after each plugin phase caught issues early. In practice, every phase passed on first attempt.
+
+### Challenges
+- **`WorkspaceEntry` boolean vs method:** The `Dirent` class uses `.isFile()` (method) while `WorkspaceEntry` uses `.isFile` (boolean property). Required mechanical find-and-replace but was easy to miss. The plan flagged this risk, which helped.
+- **Error handling for file-not-found in emit.ts:** The original code checked `(err as NodeJS.ErrnoException).code !== 'ENOENT'` but Workspace errors may not have `.code`. Changed to string matching on error messages (`!message.includes('not found') && !message.includes('ENOENT')`). This is slightly fragile but adequate for the settings.json merge use case.
+- **Path separator handling in discover.ts:** Original code used `path.sep` for splitting discovered paths. Since workspace-relative paths always use forward slashes, this was simplified to `/` splitting. Required careful review to ensure correctness.
+
+### Adherence to Standards
+- Followed TDD process: existing tests served as regression oracle; plugin-a16n parse tests rewritten before implementation.
+- No `fs/promises` imports remain in any plugin `src/` directory.
+
+## 5. Testing Phase Review
+
+- **Strategy was effective:** Using the 804 existing tests as regression oracle was the right approach for a behavioral-preserving migration.
+- **Test count grew slightly** to 807 (new tests added in Phase 0 for `toWorkspace()` and `LocalWorkspace` in models).
+- **Missing: MemoryWorkspace smoke tests.** The Phase 4 plan included "verify all plugins work with MemoryWorkspace (add smoke tests)" but this was not done. The existing tests exercise `LocalWorkspace` implicitly via string paths, proving the migration is correct, but MemoryWorkspace-based plugin tests would be the whole point of this migration and should be a follow-up task.
+
+## 6. What Went Well
+
+1. **Clean migration pattern** — `toWorkspace()` helper made the entry-point conversion trivial and consistent across all plugins
+2. **Zero regressions** — all 807 tests passed, proving behavioral preservation
+3. **Phase ordering** — starting with the smallest plugin (a16n, 11 fs calls) established the pattern before tackling larger plugins (claude 29, cursor 32)
+4. **Accurate planning** — the fs call audit and workspace method mapping table made implementation mechanical
+5. **Backward compatibility** — public APIs unchanged, no downstream impact
+
+## 7. What Could Have Been Done Differently
+
+1. **MemoryWorkspace smoke tests should have been written** — this was the stated goal of the migration but was not validated end-to-end
+2. **Error handling pattern for Workspace errors** could be standardized — the string-matching approach for ENOENT detection is fragile; a `WorkspaceError` class with error codes would be more robust
+3. **`readAgentSkillIO`/`writeAgentSkillIO` could be migrated** to accept `Workspace` directly instead of using the `ws.resolve()` bridge — this would complete the abstraction
+4. **Plugin development docs** were initially missed and required a separate update pass
+
+## 8. Key Lessons Learned
+
+### Technical
+- **Boolean properties vs methods for interfaces:** When abstracting native APIs (like Node.js `Dirent`), decide early whether interface properties should be methods or booleans. `WorkspaceEntry` chose booleans (`.isFile` instead of `.isFile()`), which is simpler but requires careful migration from method-based APIs.
+- **Forward-slash path convention for workspace:** Standardizing on `/` for all workspace-relative paths eliminates `path.sep` concerns and simplifies cross-platform support.
+- **Bridge pattern for shared utilities:** When a shared utility has many consumers and its own tests, bridging at the call site (`ws.resolve()`) is less disruptive than migrating the utility's signature.
+
+### Process
+- **Mechanical migrations benefit from clear mapping tables** — the fs-to-workspace mapping table was the single most useful planning artifact.
+- **Phase gates (build + test after each phase) provide confidence** but added minimal overhead since each phase passed on first attempt.
+- **Documentation should be updated in the same pass as implementation**, not as a separate step.
+
+## 9. Actionable Improvements for Future L3 Features
+
+1. **Add MemoryWorkspace smoke tests** as a follow-up task — this validates the migration's purpose
+2. **Consider a `WorkspaceError` class** with structured error codes to replace string-matching for error detection
+3. **Include documentation updates** in the implementation phase checklist, not as an afterthought
+4. **Migrate `readAgentSkillIO`/`writeAgentSkillIO`** to accept `Workspace` directly in a future task
+5. For future mechanical migrations, create a **migration script** if the pattern is sufficiently regular (this one had enough per-file variation to not warrant it)
