+++ memory-bank/tasks.md
@@ -1,15 +1,21 @@
 # Current Task: Migrate Bundled Plugins to Workspace Interface
 
 **Complexity Level:** 3 (Intermediate Feature)
-**Status:** Planning
+**Status:** Implementation Complete
 
 ## Task Lifecycle
 - [x] Planning complete
-- [ ] Creative phases (none required)
-- [ ] QA validation
-- [ ] Implementation
-- [ ] Reflection
+- [x] Creative phases (none required)
+- [x] QA validation
+- [x] Implementation
+- [x] Reflection
 - [ ] Archiving
+
+## Reflection Highlights
+- **What Went Well**: Clean migration pattern via `toWorkspace()`, zero regressions (807 tests), accurate planning with fs-to-workspace mapping table
+- **Challenges**: WorkspaceEntry boolean vs Dirent methods, error handling for Workspace errors (no `.code` property), path separator normalization
+- **Lessons Learned**: Boolean props vs methods decided early; forward-slash convention simplifies cross-platform; bridge pattern for shared utilities less disruptive than full migration
+- **Next Steps**: Add MemoryWorkspace smoke tests, consider WorkspaceError class, migrate readAgentSkillIO/writeAgentSkillIO to accept Workspace directly
 
 ## Context
