+++ packages/cli/src/output.ts
@@ -10,7 +10,6 @@ const ICONS: Record<WarningCode, string> = {
   'boundary-crossing': '⚡',
   'git-status-conflict': '⚠',
   'version-mismatch': '⚠',
-  'orphan-path-ref': '⚠',
 };
 
 const HINTS: Partial<Record<WarningCode, string>> = {
@@ -18,7 +17,6 @@ const HINTS: Partial<Record<WarningCode, string>> = {
   approximated: 'Behavior may differ slightly between tools',
   'boundary-crossing': 'Output file cannot be git-ignored because it is already tracked',
   'git-status-conflict': 'Source files have inconsistent git-ignore status; review manually',
-  'orphan-path-ref': 'Content references a source-format path that is not being converted; update manually',
 };
 
 /**
