+++ packages/models/test/workspace.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { resolveRoot, type Workspace } from '../src/workspace.js';
+import { resolveRoot, toWorkspace, LocalWorkspace, type Workspace } from '../src/workspace.js';
 
 describe('resolveRoot', () => {
   it('should return the string directly when given a string', () => {
@@ -18,5 +18,26 @@ describe('resolveRoot', () => {
       mkdir: async () => {},
     };
     expect(resolveRoot(mockWorkspace)).toBe('/workspace/root');
+  });
+});
+
+describe('toWorkspace', () => {
+  it('should wrap a string in a LocalWorkspace', () => {
+    const ws = toWorkspace('/project/root');
+    expect(ws).toBeInstanceOf(LocalWorkspace);
+    expect(ws.root).toBe('/project/root');
+    expect(ws.id).toBe('default');
+  });
+
+  it('should use the provided id when wrapping a string', () => {
+    const ws = toWorkspace('/project/root', 'source');
+    expect(ws).toBeInstanceOf(LocalWorkspace);
+    expect(ws.id).toBe('source');
+  });
+
+  it('should return the workspace unchanged when given a Workspace', () => {
+    const original = new LocalWorkspace('test', '/workspace/root');
+    const ws = toWorkspace(original);
+    expect(ws).toBe(original);
   });
 });
